backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
pgvector
pydantic>=2
python-multipart
passlib[bcrypt]
python-jose[cryptography]
celery[redis]
redis
minio
pytesseract
Pillow
google-generativeai>=0.5
transformers
spacy
natasha
azure-storage-blob 